Fixed bug in NetworkClient::private_read_callback; Fixed upload on pixeldrain.com
- Do not check MinVersion in servers.xml for debug builds
- Fixed inconsistent visibility of the 'Media file info' button on the Upload page

+++ Source/Core/Network/INetworkClient.h
@@ -94,8 +94,8 @@ class INetworkClient {
         virtual void clearProxy(){};
         virtual void setReferer(const std::string &str){}
         virtual void setOutputFile(const std::string &str) {}
-        virtual void setChunkOffset(double offset) {}
-        virtual void setChunkSize(double size){}
+        virtual void setChunkOffset(int64_t offset) { }
+        virtual void setChunkSize(int64_t size) { }
         virtual int getCurlResult(){ return 0; /* CURLE_OK */ }
         virtual CURL* getCurlHandle() { return nullptr;  }
         virtual void setCurlShare(CurlShare* share) {}
